+++ src/editor/mode.rs
@@ -54,9 +54,8 @@ impl<D: DataProvider> FormEditor<D> {
             {
                 match self.keybinding_paradigm() {
                     KeybindingParadigm::Helix => self.set_mode_helix(mode),
-                    KeybindingParadigm::Vim | KeybindingParadigm::Emacs => {
-                        self.set_mode_vim(mode)
-                    }
+                    KeybindingParadigm::Emacs => self.set_mode_emacs(mode),
+                    KeybindingParadigm::Vim => self.set_mode_vim(mode),
                 }
             }
             #[cfg(not(feature = "keybindings"))]
@@ -165,7 +164,8 @@ impl<D: DataProvider> FormEditor<D> {
             #[cfg(feature = "keybindings")]
             match self.keybinding_paradigm() {
                 KeybindingParadigm::Helix => self.enter_edit_mode_helix(),
-                KeybindingParadigm::Vim | KeybindingParadigm::Emacs => self.enter_edit_mode_vim(),
+                KeybindingParadigm::Emacs => self.enter_edit_mode_emacs(),
+                KeybindingParadigm::Vim => self.enter_edit_mode_vim(),
             }
             #[cfg(not(feature = "keybindings"))]
             self.enter_edit_mode_vim();
@@ -187,9 +187,8 @@ impl<D: DataProvider> FormEditor<D> {
             #[cfg(feature = "keybindings")]
             match self.keybinding_paradigm() {
                 KeybindingParadigm::Helix => self.enter_highlight_mode_helix(),
-                KeybindingParadigm::Vim | KeybindingParadigm::Emacs => {
-                    self.enter_highlight_mode_vim()
-                }
+                KeybindingParadigm::Emacs => self.enter_highlight_mode_emacs(),
+                KeybindingParadigm::Vim => self.enter_highlight_mode_vim(),
             }
             #[cfg(not(feature = "keybindings"))]
             self.enter_highlight_mode_vim();
@@ -205,9 +204,8 @@ impl<D: DataProvider> FormEditor<D> {
             #[cfg(feature = "keybindings")]
             match self.keybinding_paradigm() {
                 KeybindingParadigm::Helix => self.enter_highlight_line_mode_helix(),
-                KeybindingParadigm::Vim | KeybindingParadigm::Emacs => {
-                    self.enter_highlight_line_mode_vim()
-                }
+                KeybindingParadigm::Emacs => self.enter_highlight_line_mode_emacs(),
+                KeybindingParadigm::Vim => self.enter_highlight_line_mode_vim(),
             }
             #[cfg(not(feature = "keybindings"))]
             self.enter_highlight_line_mode_vim();
@@ -223,9 +221,8 @@ impl<D: DataProvider> FormEditor<D> {
             #[cfg(feature = "keybindings")]
             match self.keybinding_paradigm() {
                 KeybindingParadigm::Helix => self.exit_highlight_mode_helix(),
-                KeybindingParadigm::Vim | KeybindingParadigm::Emacs => {
-                    self.exit_highlight_mode_vim()
-                }
+                KeybindingParadigm::Emacs => self.exit_highlight_mode_emacs(),
+                KeybindingParadigm::Vim => self.exit_highlight_mode_vim(),
             }
             #[cfg(not(feature = "keybindings"))]
             self.exit_highlight_mode_vim();