+++ src/editor/paradigm/mod.rs
@@ -1,3 +1,4 @@
+mod emacs;
 mod helix;
 mod shared;
 mod vim;
@@ -10,9 +11,8 @@ impl<D: DataProvider> FormEditor<D> {
     pub(crate) fn apply_after_mode_change_for_paradigm(&mut self) {
         match self.behavior_state.paradigm() {
             KeybindingParadigm::Helix => self.apply_after_mode_change_helix(),
-            KeybindingParadigm::Vim | KeybindingParadigm::Emacs => {
-                self.apply_after_mode_change_vim()
-            }
+            KeybindingParadigm::Emacs => self.apply_after_mode_change_emacs(),
+            KeybindingParadigm::Vim => self.apply_after_mode_change_vim(),
         }
     }
 }