+++ src/textarea/actions/selection/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod common;
+pub(crate) mod emacs;
 pub(crate) mod helix;
 pub(crate) mod vim;