+++ src/textarea/paradigm/mod.rs
@@ -1,6 +1,8 @@
 #[cfg(feature = "keybindings")]
 mod common;
 #[cfg(feature = "keybindings")]
+mod emacs;
+#[cfg(feature = "keybindings")]
 mod helix;
 #[cfg(feature = "keybindings")]
 mod paste;